+++ titan_agent/core/multimedia/__init__.py
@@ -1,0 +1,5 @@
+"""Multimedia automation package: Video editing (FFmpeg/MoviePy) & Blender 3D rendering."""
+from .blender_engine import BlenderEngine
+from .video_engine import VideoEngine
+
+__all__ = ["BlenderEngine", "VideoEngine"]